+++ src/main/java/com/stac/daijin/domain/auth/service/RefreshTokenService.java
@@ -3,9 +3,9 @@ package com.stac.daijin.domain.auth.service;
 import com.stac.daijin.domain.auth.RefreshToken;
 import com.stac.daijin.domain.auth.presentation.dto.response.AccessTokenResponse;
 import com.stac.daijin.domain.auth.repository.RefreshTokenRepository;
-import com.stac.daijin.global.jwt.JwtProvider;
-import com.stac.daijin.global.jwt.enums.JwtType;
-import com.stac.daijin.global.jwt.exception.InvalidTokenException;
+import com.stac.daijin.global.security.jwt.JwtProvider;
+import com.stac.daijin.global.security.jwt.enums.JwtType;
+import com.stac.daijin.global.security.jwt.exception.InvalidTokenException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
