--- src/main/java/com/stac/daijin/global/jwt/enums/JwtType.java
@@ -1,6 +1,0 @@
-package com.stac.daijin.global.jwt.enums;
-
-public enum JwtType {
-    ACCESS,
-    REFRESH
-}
