+++ src/main/java/com/stac/daijin/global/security/jwt/enums/JwtType.java
@@ -1,0 +1,6 @@
+package com.stac.daijin.global.security.jwt.enums;
+
+public enum JwtType {
+    ACCESS,
+    REFRESH
+}
